requirements.txt
python-dotenv
requests
beautifulsoup4
plyer
twilio